requirements.txt
# General modules
Django==4.2.14
python-decouple
pillow
djangorestframework

#Third party modules
markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt